@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,7 +15,8 @@
  * @author     Andreas Dziemba <adziemba@web.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Marc Reimann <reimann@mediendepot-ruhr.de>
- * @copyright  2012-2019 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2022 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -24,57 +25,21 @@
 namespace MetaModels\AttributeTranslatedContentArticleBundle\Attribute;
 
 use Contao\Controller;
-use Contao\System;
+use MetaModels\AttributeTranslatedContentArticleBundle\Widgets\ContentArticleWidget;
 use MetaModels\Attribute\TranslatedReference;
-use MetaModels\IMetaModel;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use MetaModels\ITranslatedMetaModel;
 
 /**
  * This is the AttributeTranslatedContentArticle class for handling article fields.
  */
 class TranslatedContentArticle extends TranslatedReference
 {
-
-    /**
-     * The event dispatcher.
-     *
-     * @var EventDispatcherInterface
-     */
-    private $eventDispatcher;
-
     /**
      * Array of Call Ids.
      *
      * @var array
      */
     private static $arrCallIds = [];
-
-    /**
-     * Create a new instance.
-     *
-     * @param IMetaModel               $objMetaModel The MetaModel instance this attribute belongs to.
-     * @param array                    $arrData      The attribute information array.
-     * @param EventDispatcherInterface $eventDispatcher
-     */
-    public function __construct(
-        IMetaModel $objMetaModel,
-        $arrData = [],
-        EventDispatcherInterface $eventDispatcher = null
-    ) {
-        parent::__construct($objMetaModel, $arrData);
-
-        if (null === $eventDispatcher) {
-            // @codingStandardsIgnoreStart Silencing errors is discouraged
-            @\trigger_error(
-                'Event dispatcher is missing. It has to be passed in the constructor. Fallback will be dropped.',
-                E_USER_DEPRECATED
-            );
-            // @codingStandardsIgnoreEnd
-            $eventDispatcher = System::getContainer()->get('event_dispatcher');
-        }
-
-        $this->eventDispatcher = $eventDispatcher;
-    }
 
     /**
      * {@inheritdoc}
@@ -88,7 +53,7 @@ class TranslatedContentArticle extends TranslatedReference
     /**
      * {@inheritdoc}
      */
-    public function getFieldDefinition($arrOverrides = array())
+    public function getFieldDefinition($arrOverrides = [])
     {
         $arrFieldDef              = parent::getFieldDefinition($arrOverrides);
         $arrFieldDef['inputType'] = 'translatedcontentarticle';
@@ -124,7 +89,7 @@ class TranslatedContentArticle extends TranslatedReference
     /**
      * {@inheritdoc}
      */
-    public function searchForInLanguages($strPattern, $arrLanguages = array())
+    public function searchForInLanguages($strPattern, $arrLanguages = [])
     {
         // Needed to fake implement ITranslate.
         return [];
@@ -145,51 +110,82 @@ class TranslatedContentArticle extends TranslatedReference
      */
     public function getTranslatedDataFor($arrIds, $strLangCode)
     {
-        // Generate only for frontend (speeds up the backend a little)
-        if (TL_MODE == 'BE') {
-            return [];
-        }
+        $strTable       = $this->getMetaModel()->getTableName();
+        $strColumn      = $this->getColName();
+        $model          = $this->getMetaModel();
+        $arrData        = [];
+        $contentArticle = new ContentArticleWidget();
+        $rootTable      = $contentArticle->getRootMetaModelTable($strTable);
 
-        $strTable            = $this->getMetaModel()->getTableName();
-        $strColumn           = $this->getColName();
-        $strLanguage         = $this->getMetaModel()->isTranslated() ? $strLangCode : '-';
-        $strFallbackLanguage = $this->getMetaModel()->isTranslated()
-            ? $this->getMetaModel()->getFallbackLanguage()
-            : '-';
-        $arrData             = [];
+        if ($this->getMetaModel() instanceof ITranslatedMetaModel) {
+            $strLanguage         = $strLangCode;
+            $strFallbackLanguage = $model->getMainLanguage();
+        } else {
+            $strLanguage         = $model->isTranslated() ? $strLangCode : '';
+            $strFallbackLanguage = $model->isTranslated() ? $model->getFallbackLanguage() : '';
+        }
 
         foreach ($arrIds as $intId) {
             // Continue if it's a recursive call
             $strCallId = $strTable . '_' . $strColumn . '_' . $strLanguage . '_' . $intId;
             if (isset(static::$arrCallIds[$strCallId])) {
-                $arrData[$intId]['value'] = sprintf('RECURSION: %s', $strCallId);
+                $arrData[$intId]['value'] = \sprintf('RECURSION: %s', $strCallId);
                 continue;
             }
             static::$arrCallIds[$strCallId] = true;
 
-            $objContent         = \ContentModel::findPublishedByPidAndTable($intId, $strTable);
-            $arrContent         = [];
-            $arrContentFallback = [];
-
-            if ($objContent !== null) {
-                while ($objContent->next()) {
-                    if ($objContent->mm_slot == $strColumn && $objContent->mm_lang == $strLanguage) {
-                        $arrContent[] = Controller::getContentElement($objContent->current());
-                    } elseif ($objContent->mm_slot == $strColumn
-                        && $strLanguage != $strFallbackLanguage
-                        && $objContent->mm_lang == $strFallbackLanguage
-                    ) {
-                        $arrContentFallback[] = Controller::getContentElement($objContent->current());
+            // Generate list for backend.
+            if (TL_MODE == 'BE') {
+                $elements = $contentArticle->getContentTypesByRecordId($intId, $rootTable, $strColumn, $strLanguage);
+                $content  = '';
+                if (count($elements)) {
+                    $content .= '<ul class="elements_container">';
+                    foreach ((array) $elements as $element) {
+                        // @codingStandardsIgnoreStart - one line template.
+                        $content .= \sprintf(
+                            '<li><div class="cte_type%s"><img src="system/themes/flexible/icons/%s.svg" width="16" height="16"> %s</div></li>',
+                            $element['isInvisible'] ? ' unpublished' : ' published',
+                            $element['isInvisible'] ? 'invisible' : 'visible',
+                            $element['name']
+                        );
+                        // @codingStandardsIgnoreEnd
                     }
+                    $content .= '</ul>';
+                }
+
+                if (!empty($content)) {
+                    $arrData[$intId]['value'] = [$content];
+                } else {
+                    $arrData[$intId]['value'] = [];
                 }
             }
 
-            if (!empty($arrContent)) {
-                $arrData[$intId]['value'] = $arrContent;
-            } elseif (!empty($arrContentFallback)) {
-                $arrData[$intId]['value'] = $arrContentFallback;
-            } else {
-                $arrData[$intId]['value'] = [];
+            // Generate output for frontend.
+            if (TL_MODE == 'FE') {
+                $objContent         = \ContentModel::findPublishedByPidAndTable($intId, $strTable);
+                $arrContent         = [];
+                $arrContentFallback = [];
+
+                if ($objContent !== null) {
+                    while ($objContent->next()) {
+                        if ($objContent->mm_slot == $strColumn && $objContent->mm_lang == $strLanguage) {
+                            $arrContent[] = Controller::getContentElement($objContent->current());
+                        } elseif ($objContent->mm_slot == $strColumn
+                                  && $strLanguage != $strFallbackLanguage
+                                  && $objContent->mm_lang == $strFallbackLanguage
+                        ) {
+                            $arrContentFallback[] = Controller::getContentElement($objContent->current());
+                        }
+                    }
+                }
+
+                if (!empty($arrContent)) {
+                    $arrData[$intId]['value'] = $arrContent;
+                } elseif (!empty($arrContentFallback)) {
+                    $arrData[$intId]['value'] = $arrContentFallback;
+                } else {
+                    $arrData[$intId]['value'] = [];
+                }
             }
 
             unset(static::$arrCallIds[$strCallId]);
