@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2021 The MetaModels team.
+ * (c) 2012-2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,7 +15,7 @@
  * @author     Andreas Dziemba <adziemba@web.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2021 The MetaModels team.
+ * @copyright  2012-2022 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -23,10 +23,8 @@
 namespace MetaModels\AttributeTranslatedContentArticleBundle\Attribute;
 
 use Doctrine\DBAL\Connection;
-use MetaModels\Attribute\AbstractAttributeTypeFactory;
 use MetaModels\Attribute\AbstractSimpleAttributeTypeFactory;
 use MetaModels\Helper\TableManipulator;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * Attribute type factory for article attributes.
@@ -34,26 +32,17 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 class AttributeTypeFactory extends AbstractSimpleAttributeTypeFactory
 {
     /**
-     * Event dispatcher.
-     *
-     * @var EventDispatcherInterface
-     */
-    private $eventDispatcher;
-
-    /**
      * {@inheritDoc}
      */
     public function __construct(
-        EventDispatcherInterface $eventDispatcher,
         Connection $connection,
         TableManipulator $tableManipulator
     ) {
         parent::__construct($connection, $tableManipulator);
 
-        $this->typeName        = 'translatedcontentarticle';
-        $this->typeIcon        = 'bundles/metamodelsattributetranslatedcontentarticle/article.png';
-        $this->typeClass       = TranslatedContentArticle::class;
-        $this->eventDispatcher = $eventDispatcher;
+        $this->typeName  = 'translatedcontentarticle';
+        $this->typeIcon  = 'bundles/metamodelsattributetranslatedcontentarticle/article.png';
+        $this->typeClass = TranslatedContentArticle::class;
     }
 
     /**
@@ -61,6 +50,6 @@ class AttributeTypeFactory extends AbstractSimpleAttributeTypeFactory
      */
     public function createInstance($information, $metaModel)
     {
-        return new $this->typeClass($metaModel, $information, $this->eventDispatcher);
+        return new $this->typeClass($metaModel, $information);
     }
 }
