@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_translatedcontentarticle.
  *
- * (c) 2012-2021 The MetaModels team.
+ * (c) 2012-2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,7 +15,7 @@
  * @author     Andreas Dziemba <adziemba@web.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
- * @copyright  2012-2021 The MetaModels team.
+ * @copyright  2012-2022 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_translatedcontentarticle/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -23,21 +23,23 @@
 namespace MetaModels\AttributeTranslatedContentArticleBundle\Widgets;
 
 use Contao\CoreBundle\Framework\Adapter;
+use Contao\Environment;
 use Contao\Input;
 use Contao\System;
-use Contao\Widget;
 use ContaoCommunityAlliance\DcGeneral\Contao\Compatibility\DcCompat;
+use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\ContaoBackendViewTemplate;
+use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Widget\AbstractWidget;
 use ContaoCommunityAlliance\DcGeneral\Data\MultiLanguageDataProviderInterface;
 use Doctrine\DBAL\Connection;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * Class ArticleWidget
  *
  * @package MetaModels\AttributeTranslatedContentArticleBundle\Widgets
  */
-class ContentArticleWidget extends Widget
+class ContentArticleWidget extends AbstractWidget
 {
-
     /**
      * Submit user input.
      *
@@ -57,7 +59,7 @@ class ContentArticleWidget extends Widget
      *
      * @var string
      */
-    protected $strTemplate = 'be_widget';
+    protected $subTemplate = 'widget_contentarticle';
 
     /**
      * Flag if the current entry has an id.
@@ -92,7 +94,6 @@ class ContentArticleWidget extends Widget
      *
      * @var \ContaoCommunityAlliance\DcGeneral\Contao\Compatibility\DcCompat|null
      */
-    private $dcCompat;
 
     /**
      * Check if we have an id, if not set a flag.
@@ -104,18 +105,101 @@ class ContentArticleWidget extends Widget
         $arrAttributes = null,
         DcCompat $dcCompat = null,
         Connection $connection = null,
-        Adapter $input = null
+        Adapter $input = null,
+        TranslatorInterface $translator = null
     ) {
-        $this->dcCompat   = $dcCompat;
-        $this->connection = ($connection ?? System::getContainer()->get('database_connection'));
-        $this->input      = (
-            $input ?? System::getContainer()->get('contao.framework')->getAdapter(Input::class)
-        );
+        if (null === $connection) {
+            // @codingStandardsIgnoreStart
+            @trigger_error(
+                'Connection is missing. It has to be passed in the constructor. Fallback will be dropped.',
+                E_USER_DEPRECATED
+            );
+            // @codingStandardsIgnoreEnd
+            $connection = System::getContainer()->get('database_connection');
+        }
+        $this->connection = $connection;
 
-        parent::__construct($arrAttributes);
+        if (null === $input) {
+            // @codingStandardsIgnoreStart
+            @trigger_error(
+                'Input adapter is missing. It has to be passed in the constructor. Fallback will be dropped.',
+                E_USER_DEPRECATED
+            );
+            // @codingStandardsIgnoreEnd
+            $input = System::getContainer()->get('contao.framework')->getAdapter(Input::class);
+        }
+        $this->input = $input;
+
+        if (null === $translator) {
+            // @codingStandardsIgnoreStart
+            @trigger_error(
+                'Translator is missing. It has to be passed in the constructor. Fallback will be dropped.',
+                E_USER_DEPRECATED
+            );
+            // @codingStandardsIgnoreEnd
+            $translator = System::getContainer()->get('translator');
+        }
+        $this->translator = $translator;
+
+        parent::__construct($arrAttributes, $dcCompat);
 
         $currentID        = $this->input->get('id');
         $this->hasEmptyId = empty($currentID);
+    }
+
+    /**
+     * Set an object property
+     *
+     * @param string $strKey   The property name.
+     * @param mixed  $varValue The property value.
+     *
+     * @return void
+     */
+    public function __set($strKey, $varValue)
+    {
+        switch ($strKey) {
+            case 'subTemplate':
+                $this->subTemplate = $varValue;
+                break;
+            default:
+                parent::__set($strKey, $varValue);
+                break;
+        }
+    }
+
+    /**
+     * Return an object property
+     *
+     * @param string $strKey The property name.
+     *
+     * @return string The property value
+     */
+    public function __get($strKey)
+    {
+        switch ($strKey) {
+            case 'subTemplate':
+                return $this->subTemplate;
+            default:
+        }
+
+        return parent::__get($strKey);
+    }
+
+    /**
+     * Check whether an object property exists
+     *
+     * @param string $strKey The property name.
+     *
+     * @return boolean True if the property exists
+     */
+    public function __isset($strKey)
+    {
+        switch ($strKey) {
+            case 'subTemplate':
+                return isset($this->subTemplate);
+            default:
+                return parent::__get($strKey);
+        }
     }
 
     /**
@@ -129,91 +213,88 @@ class ContentArticleWidget extends Widget
      */
     public function generate()
     {
-        // Update the language.
-        /** @var \MetaModels\DcGeneral\Data\Driver $dataProvider */
-        $dataProvider = $this->dcCompat->getEnvironment()->getDataProvider();
-        if ($dataProvider instanceof MultiLanguageDataProviderInterface) {
-            $currentLang = $dataProvider->getCurrentLanguage();
+        // Retrieve current language.
+        $currentLang = '';
+        if (Environment::get('isAjaxRequest')) {
+            $currentLang = Input::post('lang');
         } else {
-            $currentLang = null;
-        }
-        $this->lang = ($currentLang) ?: '-';
-
-        if (!empty($GLOBALS['TL_LANG']['MSC']['edit'])) {
-            $edit = $GLOBALS['TL_LANG']['MSC']['edit'];
-        } else {
-            $edit = 'Edit';
+            $dataProvider = $this->getEnvironment()->getDataProvider();
+            if ($dataProvider instanceof MultiLanguageDataProviderInterface) {
+                $currentLang = $dataProvider->getCurrentLanguage();
+            }
         }
 
-        // If we have no id, we get some trouble with the modal. So we disabled the button.
-        if ($this->hasEmptyId) {
-            return sprintf(
-                '<p class="tl_help tl_tip">%s</p>' .
-                '<button type="button" name="%s" class="tl_submit" disabled>%s</button>',
-                $GLOBALS['TL_LANG']['attribute_contentarticle']['missing_id'],
-                $this->name,
-                $edit
-            );
-        }
+        $rootTable = $this->getRootMetaModelTable($this->strTable);
 
         $strQuery = http_build_query([
-            'do'          => 'metamodel_' . $this->getRootMetaModelTable($this->strTable) ?: 'table_not_found',
-            'table'       => 'tl_content',
-            'ptable'      => $this->strTable,
-            'id'          => $this->currentRecord,
-            'mid'         => $this->currentRecord,
-            'slot'        => $this->strName,
-            'lang'        => $this->lang,
-            'popup'       => 1,
-            'nb'          => 1,
-            'langSupport' => 1,
-            'rt'          => REQUEST_TOKEN,
-        ]);
+                                         'do'          => 'metamodel_' . ($rootTable ?: 'table_not_found'),
+                                         'table'       => 'tl_content',
+                                         'ptable'      => $this->strTable,
+                                         'id'          => $this->currentRecord,
+                                         'mid'         => $this->currentRecord,
+                                         'slot'        => $this->strName,
+                                         'lang'        => $currentLang,
+                                         'popup'       => 1,
+                                         'nb'          => 1,
+                                         'langSupport' => 1,
+                                         'rt'          => REQUEST_TOKEN,
+                                     ]);
 
-        return sprintf(
-            '<div><p><a href="%s" class="tl_submit" onclick="%s">%s</a></p></div>',
-            'contao?' . $strQuery,
-            'Backend.openModalIframe({width:850,title:\'' . $this->strLabel . '\',url:this.href});return false',
-            $edit
-        );
+        $contentElements =
+            $this->getContentTypesByRecordId($this->currentRecord, $rootTable, $this->strName, $currentLang);
+
+        $content = (new ContaoBackendViewTemplate($this->subTemplate))
+            ->setTranslator($this->getEnvironment()->getTranslator())
+            ->set('name', $this->strName)
+            ->set('id', $this->strId)
+            ->set('label', $this->label)
+            ->set('readonly', $this->readonly)
+            ->set('hasEmptyId', $this->hasEmptyId)
+            ->set('link', 'contao?' . $strQuery)
+            ->set('elements', $contentElements)
+            ->set('lang', $currentLang)
+            ->parse();
+
+        return !Environment::get('isAjaxRequest') ? '<div>' . $content . '</div>' : $content;
     }
 
     /**
      * Get the RootMetaModelTable.
      *
-     * @param string $strTable Table name to Check.
+     * @param string $tableName Table name to Check.
      *
      * @return bool|string Returns RootMetaModelTable.
      *
      * @throws \Exception Throws an Exception.
      */
-    private function getRootMetaModelTable($strTable)
+    public function getRootMetaModelTable(string $tableName)
     {
-        $arrTables = [];
-        $objTables = \Database::getInstance()->execute('
-                SELECT tableName, d.renderType, d.ptable
-                FROM tl_metamodel AS m
-                JOIN tl_metamodel_dca AS d
-                ON m.id = d.pid
-            ');
+        $tables = [];
 
-        while ($objTables->next()) {
-            $arrTables[$objTables->tableName] = [
-                'renderType' => $objTables->renderType,
-                'ptable'     => $objTables->ptable,
+        $statement = $this->connection
+            ->createQueryBuilder()
+            ->select('t.tableName, d.renderType, d.ptable')
+            ->from('tl_metamodel', 't')
+            ->leftJoin('t', 'tl_metamodel_dca', 'd', '(t.id=d.pid)')
+            ->execute();
+
+        while ($row = $statement->fetchAssociative()) {
+            $tables[$row['tableName']] = [
+                'renderType' => $row['renderType'],
+                'ptable'     => $row['ptable'],
             ];
         }
 
-        $getTable = function ($strTable) use (&$getTable, $arrTables) {
-            if (!isset($arrTables[$strTable])) {
+        $getTable = function ($tableName) use (&$getTable, $tables) {
+            if (!isset($tables[$tableName])) {
                 return false;
             }
 
-            $arrTable = $arrTables[$strTable];
+            $arrTable = $tables[$tableName];
 
             switch ($arrTable['renderType']) {
                 case 'standalone':
-                    return $strTable;
+                    return $tableName;
 
                 case 'ctable':
                     return $getTable($arrTable['ptable']);
@@ -223,6 +304,56 @@ class ContentArticleWidget extends Widget
             }
         };
 
-        return $getTable($strTable);
+        return $getTable($tableName);
+    }
+
+
+    /**
+     * Retrieve all content elements of this item as parent.
+     *
+     * @param int|null    $recordId    The record id.
+     * @param string      $ptableName  The name of parent tab.
+     * @param string      $slotName    The name of slot.
+     * @param string|null $currentLang The current language.
+     *
+     * @return array Returns array with content elements.
+     */
+    public function getContentTypesByRecordId(
+        ?int $recordId,
+        string $ptableName,
+        string $slotName,
+        ?string $currentLang
+    ): array {
+        $contentElements = [];
+
+        if (empty($recordId) || empty($ptableName)) {
+            return $contentElements;
+        }
+
+        $statement = $this->connection
+            ->createQueryBuilder()
+            ->select('t.type, t.invisible, t.start, t.stop')
+            ->from('tl_content', 't')
+            ->where('t.pid=:pid')
+            ->andWhere('t.ptable=:ptable')
+            ->andWhere('t.mm_slot=:slot')
+            ->andWhere('t.mm_lang=:lang')
+            ->orderBy('t.sorting')
+            ->setParameter('pid', $recordId)
+            ->setParameter('ptable', $ptableName)
+            ->setParameter('slot', $slotName)
+            ->setParameter('lang', $currentLang)
+            ->execute();
+
+        while ($row = $statement->fetchAssociative()) {
+            $contentElements[] = [
+                'name'        => $this->translator->trans('CTE.' . $row['type'] . '.0', [], 'contao_default'),
+                'isInvisible' => $row['invisible']
+                                 || ($row['start'] && $row['start'] > time())
+                                 || ($row['stop'] && $row['stop'] <= time())
+            ];
+        }
+
+        return $contentElements;
     }
 }
