@@ -5,7 +5,7 @@ declare(strict_types=1);
 /**
  * This file is part of MetaModels/attribute_contentarticle.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2022 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,6 +15,7 @@ declare(strict_types=1);
  * @package    MetaModels
  * @subpackage AttributeContentArticle
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
+ * @author     Sven Baumann <baumann.sv@gmail.com>
  */
 
 namespace Widgets;
@@ -25,6 +26,7 @@ use ContaoCommunityAlliance\DcGeneral\Contao\Compatibility\DcCompat;
 use Doctrine\DBAL\Connection;
 use MetaModels\AttributeTranslatedContentArticleBundle\Widgets\ContentArticleWidget;
 use PHPUnit\Framework\TestCase;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * @covers \MetaModels\AttributeTranslatedContentArticleBundle\Widgets\ContentArticleWidget
@@ -50,8 +52,11 @@ class ContentArticleWidgetTest extends TestCase
             ->method('get')
             ->willReturn(1);
 
+        $translator = $this->getMockBuilder(TranslatorInterface::class)
+                            ->getMock();
+
         $widget = $this->getMockBuilder(ContentArticleWidget::class)
-                       ->setConstructorArgs([null, $dcCompat, $connection, $input])
+                       ->setConstructorArgs([null, $dcCompat, $connection, $input, $translator])
                        ->setMethods(['import'])
                        ->getMock();
 
